MythHTTPParser: Fix qMax usage on some platforms

+++ mythtv/libs/libmythbase/http/mythhttpparser.cpp
@@ -131,7 +131,7 @@ bool MythHTTPParser::Read(QTcpSocket* Socket, bool& Ready)
     {
         int64_t want = m_contentLength - m_content->size();
         int64_t have = Socket->bytesAvailable();
-        m_content->append(Socket->read(qMax(want, qMax(HTTP_CHUNKSIZE, have))));
+        m_content->append(Socket->read(qMax(want, qMax(static_cast<int64_t>(HTTP_CHUNKSIZE), have))));
     }
 
     // Need more data...
